refactor: Create separate package for audio-related code
- Create new package – `ai.elimu.util.audio` – where we will put audio-related and Text-to-Speech related utility code.
