+++ src/main/java/ai/elimu/rest/v2/crowdsource/AudioContributionsRestController.java
@@ -12,8 +12,8 @@ import ai.elimu.model.enums.Platform;
 import ai.elimu.model.enums.content.AudioFormat;
 import ai.elimu.model.v2.gson.content.WordGson;
 import ai.elimu.rest.v2.JpaToGsonConverter;
-import ai.elimu.util.AudioMetadataExtractionHelper;
-import ai.elimu.util.CrowdsourceHelper;
+import ai.elimu.util.audio.AudioMetadataExtractionHelper;
+import ai.elimu.util.audio.CrowdsourceHelper;
 import com.google.gson.Gson;
 import java.io.File;
 import java.util.ArrayList;
